backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
Pillow
requests
django-cors-headers
djangorestframework-authtoken